requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Werkzeug==2.3.7

# Document processing
PyPDF2==3.0.1
python-docx==0.8.11
python-pptx==0.6.21  # For PowerPoint files (future enhancement)
openpyxl==3.1.2      # For Excel files (future enhancement)

# Development tools
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
isort==5.12.0

# Additional utilities
python-dotenv==1.0.0
bcrypt==4.0.1
